gcs-sync-job/requirements.txt
google-genai>=1.49.0
google-cloud-storage>=2.14.0
